Img2Story/requirements.txt
python-dotenv==1.0.0
transformers==4.33.1
torch==2.0.1
langchain==0.0.304
streamlit==1.27.1
openai==0.28.1